fix(async functions): properly entangle for the value and prevent infinite re-rendering
